listener: extract active tcp listener out of connection handler impl
