+++ source/server/BUILD
@@ -61,6 +61,7 @@ envoy_cc_library(
 envoy_cc_library(
     name = "connection_handler_lib",
     deps = [
+        ":active_tcp_listener",
         ":active_udp_listener",
         ":connection_handler_impl",
     ],
@@ -70,7 +71,34 @@ envoy_cc_library(
     name = "connection_handler_impl",
     srcs = ["connection_handler_impl.cc"],
     hdrs = [
+        "active_tcp_listener.h",
         "connection_handler_impl.h",
+    ],
+    deps = [
+        "//include/envoy/common:time_interface",
+        "//include/envoy/event:deferred_deletable",
+        "//include/envoy/event:dispatcher_interface",
+        "//include/envoy/network:connection_handler_interface",
+        "//include/envoy/network:connection_interface",
+        "//include/envoy/network:filter_interface",
+        "//include/envoy/network:listen_socket_interface",
+        "//include/envoy/network:listener_interface",
+        "//include/envoy/server:active_udp_listener_config_interface",
+        "//include/envoy/server:listener_manager_interface",
+        "//include/envoy/stats:timespan_interface",
+        "//source/common/common:linked_object",
+        "//source/common/common:non_copyable",
+        "//source/common/event:deferred_task",
+        "//source/common/network:connection_lib",
+        "//source/common/stream_info:stream_info_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "active_tcp_listener",
+    srcs = ["active_tcp_listener.cc"],
+    hdrs = [
+        "active_tcp_listener.h",
     ],
     deps = [
         "//include/envoy/common:time_interface",
@@ -79,7 +107,6 @@ envoy_cc_library(
         "//include/envoy/event:timer_interface",
         "//include/envoy/network:connection_handler_interface",
         "//include/envoy/network:connection_interface",
-        "//include/envoy/network:exception_interface",
         "//include/envoy/network:filter_interface",
         "//include/envoy/network:listen_socket_interface",
         "//include/envoy/network:listener_interface",
@@ -93,6 +120,7 @@ envoy_cc_library(
         "//source/common/stats:timespan_lib",
         "//source/common/stream_info:stream_info_lib",
         "//source/extensions/transport_sockets:well_known_names",
+        "//source/server:connection_handler_impl",
     ],
 )
